rename module to vertex_rtree
