+++ rust/routee-compass/src/plugin/input/default/mod.rs
@@ -1,2 +1,2 @@
 pub mod grid_search;
-pub mod rtree;
+pub mod vertex_rtree;
