+++ rust/routee-compass/src/plugin/input/default/vertex_rtree/plugin.rs
